added base for relative node scaling

+++ core/layoutEngine/ILayoutEngine.hpp
@@ -10,7 +10,7 @@ class ILayoutEngine
 {
 public:
     // Normal
-    virtual glm::ivec2 process(const AbstractNodePtr& parent) = 0;
+    virtual glm::vec2 process(const AbstractNodePtr& parent) = 0;
 };
 
 using ILayoutEnginePtr = std::shared_ptr<ILayoutEngine>;
